remove legacy compat, refactor
