--- src/main/java/com/tacz/guns/compat/iris/newly/pbr/PBRRegister.java
@@ -1,12 +1,0 @@
-package com.tacz.guns.compat.iris.newly.pbr;
-
-import com.tacz.guns.client.resource.texture.FilePackTexture;
-import com.tacz.guns.client.resource.texture.ZipPackTexture;
-import net.irisshaders.iris.texture.pbr.loader.PBRTextureLoaderRegistry;
-
-public class PBRRegister {
-    public static void registerPBRLoader() {
-        PBRTextureLoaderRegistry.INSTANCE.register(FilePackTexture.class, new FilePackTexturePBRLoader());
-        PBRTextureLoaderRegistry.INSTANCE.register(ZipPackTexture.class, new ZipPackTexturePBRLoader());
-    }
-}
