+++ src/main/java/com/tacz/guns/compat/iris/pbr/PBRRegister.java
@@ -1,8 +1,9 @@
-package com.tacz.guns.compat.iris.legacy.pbr;
+package com.tacz.guns.compat.iris.pbr;
 
 import com.tacz.guns.client.resource.texture.FilePackTexture;
 import com.tacz.guns.client.resource.texture.ZipPackTexture;
-import net.coderbot.iris.texture.pbr.loader.PBRTextureLoaderRegistry;
+
+import net.irisshaders.iris.texture.pbr.loader.PBRTextureLoaderRegistry;
 
 public class PBRRegister {
     public static void registerPBRLoader() {
