+++ src/main/resources/fabric.mod.json
@@ -56,6 +56,9 @@
 		"porting_lib_entity": ">=2.3.4+1.20.1",
 		"mkb": ">=1.2.0"
 	},
+	"breaks": {
+		"iris": "<1.7.0"
+	},
 	"custom": {
 		"loom:injected_interfaces": {
 			"net/minecraft/class_276": [
